fix typo
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@255451 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Zend/zend_operators.h
@@ -336,7 +336,7 @@ ZEND_API void zend_compare_arrays(zval *result, zval *a1, zval *a2 TSRMLS_DC);
 ZEND_API void zend_compare_objects(zval *result, zval *o1, zval *o2 TSRMLS_DC);
 
 ZEND_API int zend_atoi(const char *str, int str_len);
-ZEND_API int zend_atol(const char *str, int str_len);
+ZEND_API long zend_atol(const char *str, int str_len);
 
 ZEND_API void zend_locale_sprintf_double(zval *op ZEND_FILE_LINE_DC);
 END_EXTERN_C()
